requirements.txt
google-generativeai
semgrep
